+++ zarr_creator/read_source.py
@@ -35,10 +35,35 @@ def read_level_type_data(t_analysis: datetime.datetime, level_type: str) -> xr.D
         if "cfName" in ds[var_name].attrs:
             ds[var_name].attrs["standard_name"] = ds[var_name].attrs["cfName"]
 
-    if level_type == "heightAboveGround":
-        # u-wind @ 10m and 100m are given as their own variables... same for v-wind
-        ds = _merge_special_fields(ds)
+        # https://www.nco.ncep.noaa.gov/pmb/docs/grib2/grib2_doc/grib2_table4-2-0-4.shtml
+        if var_name == "swavr":
+            ds[var_name].attrs["standard_name"] = "surface_downwelling_shortwave_flux"
+            ds[var_name].attrs["long_name"] = "Surface downwelling shortwave flux"
+            ds[var_name].attrs["units"] = "W m-2"
+        elif var_name == "swavr_accum":
+            ds[var_name].attrs[
+                "long_name"
+            ] = "Accumulated surface downwelling shortwave flux"
+            ds[var_name].attrs["units"] = "J m-2"
+        # https://www.nco.ncep.noaa.gov/pmb/docs/grib2/grib2_doc/grib2_table4-2-0-5.shtml
+        elif var_name == "lwavr":
+            # the parameterCategory is 5 (radiation) and the parameterNumber is
+            # 4 ("Upward Longwave Radiation Flux"), but looking at the data
+            # (near-zero where there is cloud, negative otherwise)
+            # it seems to be net downward longwave flux, not upward flux as the
+            # parameter name suggests.
+            ds[var_name].attrs["standard_name"] = "surface_net_downward_longwave_flux"
+            ds[var_name].attrs["long_name"] = "Surface downwelling longwave flux"
+            ds[var_name].attrs["units"] = "W m-2"
+        elif var_name == "lwavr_accum":
+            # the accumulated values however are positive where there is no
+            # cloud, which suggests to me to the flux is upwards
+            ds[var_name].attrs[
+                "long_name"
+            ] = "Accumulated surface net upward longwave flux"
+            ds[var_name].attrs["units"] = "J m-2"
 
+    if level_type == "heightAboveGround":
         # land-sea mask is given for each timestep even though it doesn't
         # change, let's remove the time dimension
         ds["lsm"] = ds.isel(time=0).lsm
@@ -131,67 +156,6 @@ def _add_projection_info(ds):
 
     for var_name in ds.data_vars:
         ds[var_name].attrs["grid_mapping"] = PROJECTION_IDENTIFIER
-
-
-def _merge_special_fields(ds):
-    # u-wind component in general has paramter-id 131, but at 10m altitude is
-    # parameter 165 and at 100m altitude is parameter 228246:
-    #    shortName   paramId     level
-    #    10u         165         10
-    #    u           131         50
-    #    100u        228246      100
-    #    u           131         150
-    #    u           131         250
-    #    u           131         350
-    #    u           131         450
-    # same for v-wind component (in general paramId 132, 166 @ 10m, 228247 @ 100m):
-    #    shortName   paramId     level
-    #    10v         166         10
-    #    v           132         50
-    #    100v        228247      100
-    #    v           132         150
-    #    v           132         250
-    #    v           132         350
-    #    v           132         450
-
-    # this means that `u` and `v` actually have nan values at 10m and 100m
-    # altitudes, which we should replace with the values from `10u` and `10v`
-    # and `100u` and `100v` respectively
-
-    ds_copy = ds.copy()
-
-    special_params = {
-        "u": {
-            "10u": 10,
-            "100u": 100,
-        },
-        "v": {
-            "10v": 10,
-            "100v": 100,
-        },
-        "t": {
-            "2t": 2,
-        },
-    }
-
-    data_arrays = {}
-    for true_param, special_param_info in special_params.items():
-        da = ds[true_param]
-
-        for special_param, level in special_param_info.items():
-            keep_levels = [lev for lev in da.level.values if lev != level]
-            da_subset = da.sel(level=keep_levels)
-
-            da_special = ds[special_param]
-            da_special["level"] = level
-
-            da = xr.concat([da_subset, da_special], dim="level")
-        data_arrays[true_param] = da
-
-    ds_copy = xr.merge(
-        list(data_arrays.values()) + [ds.drop_vars(list(special_params.keys()))]
-    )
-    return ds_copy
 
 
 def merge_level_specific_params(ds, true_param, level, short_name):